requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# GUI
PySide6>=6.6.0

# Database
sqlalchemy>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.25.0  # Asynchronní HTTP klient pro Vehicle Decoder Engine

# Desktop webview (pro admin launcher)
pywebview>=4.4.1

# Validation
pydantic[email]>=2.5.0

# Security - password hashing and JWT
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Web automation (for VIN decoder)
selenium>=4.15.0
webdriver-manager>=4.0.0

# File watching
watchdog>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.0
reportlab>=4.0.0

# Image processing
Pillow>=10.0.0

# Process management
psutil>=5.9.0

# System tray
pystray>=0.19.0

# Email
aiosmtplib>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Audio (optional - for voice features)
# SpeechRecognition>=3.10.0
# pyttsx3>=2.90
